+++ server/index.ts
@@ -26,7 +26,7 @@ if (require.main === module) {
   app.listen(PORT, () => {
     console.log('server started at http://localhost:' + PORT);
     authDbConnection();
-    // syncModels();
+    syncModels();
   });
 }
