+++ server/models/stuff.sql
@@ -1,0 +1,11 @@
+insert into test_case values(1, 'testcase1', NOW(), NOW(), 'step', 'price', 'dist', 'type1');
+
+insert into account values(1, 'chris', 'chris@gmail.com', 'pw', 'apikey');
+
+insert into controller_property values(1, 'type1', 'problemform', 'model1', 6, 100);
+
+insert into kpi values (1, 2, 3, 4, 5, 6, 7);
+
+insert into test_case values (1, 'test1', NOW(), NOW(), 'control', 'price', 'dist', 'building1');
+
+insert into result values (1, 1, 1, 1, 1, NOW(), true);
